# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: C100 System API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /governance/documents:
    get:
      operationId: listGovernanceDocuments
      tags: [GovernanceDocs]
      summary: List governance documents
      responses:
        "200":
          description: List of governance documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createGovernanceDocument
      tags: [GovernanceDocs]
      summary: Create a governance document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernanceDocumentCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"
  /governance/documents/{id}:
    get:
      operationId: getGovernanceDocument
      tags: [GovernanceDocs]
      summary: Get a governance document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Governance document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateGovernanceDocument
      tags: [GovernanceDocs]
      summary: Update a governance document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernanceDocumentCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"
  /governance/documents/{id}/publish:
    post:
      operationId: publishGovernanceDocument
      tags: [GovernanceDocs]
      summary: Publish a governance document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"
  /governance/documents/{id}/supersede:
    post:
      operationId: supersedeGovernanceDocument
      tags: [GovernanceDocs]
      summary: Supersede a governance document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Superseded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"
  /governance/documents/{id}/archive:
    post:
      operationId: archiveGovernanceDocument
      tags: [GovernanceDocs]
      summary: Archive a governance document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"
  /governance/documents/{id}/restore:
    post:
      operationId: restoreGovernanceDocument
      tags: [GovernanceDocs]
      summary: Restore an archived governance document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceDocument"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Secretary ───────────────────────────────────────────────────────────────
...
  /conduct/records:
    get:
      operationId: listConductRecords
      tags: [Conduct]
      summary: List conduct records (confidential)
      responses:
        "200":
          description: Conduct records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConductRecord"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createConductRecord
      tags: [Conduct]
      summary: Create a conduct record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConductRecordCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConductRecord"
  /conduct/records/{id}:
    patch:
      operationId: patchConductRecord
      tags: [Conduct]
      summary: Update a conduct record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConductRecordCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConductRecord"
  /conduct/records/{id}/resolve:
    post:
      operationId: resolveConductRecord
      tags: [Conduct]
      summary: Resolve a conduct record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConductRecordResolveBody"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConductRecord"
  /conduct/records/{id}/archive:
    post:
      operationId: archiveConductRecord
      tags: [Conduct]
      summary: Archive a conduct record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConductRecord"

  # ── Procedure ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
...
  responses:
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    GovernanceDocument:
      type: object
      required: [id, title, category, versionLabel, effectiveDate, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        versionLabel:
          type: string
        effectiveDate:
          type: string
        approvalDate:
          type: [string, 'null']
        status:
          type: string
        notes:
          type: [string, 'null']
        originalFilename:
          type: [string, 'null']
        mimeType:
          type: [string, 'null']
        fileSizeBytes:
          type: [integer, 'null']
        storageKey:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GovernanceDocumentCreate:
      type: object
      required: [title, category, versionLabel, effectiveDate]
      properties:
        title:
          type: string
        category:
          type: string
        versionLabel:
          type: string
        effectiveDate:
          type: string
        approvalDate:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        originalFilename:
          type: [string, 'null']
        mimeType:
          type: [string, 'null']
        fileSizeBytes:
          type: [integer, 'null']
        storageKey:
          type: [string, 'null']

    # ── Secretary schemas ────────────────────────────────────────────────────
...
    ConductRecord:
      type: object
      required: [id, recordType, reportDate, summary, privateDetails, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        recordType:
          type: string
        reportDate:
          type: string
        memberId:
          type: [integer, 'null']
        eventId:
          type: [integer, 'null']
        summary:
          type: string
        privateDetails:
          type: string
        status:
          type: string
        resolution:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConductRecordCreate:
      type: object
      required: [recordType, reportDate, summary]
      properties:
        recordType:
          type: string
        reportDate:
          type: string
        memberId:
          type: [integer, 'null']
        eventId:
          type: [integer, 'null']
        summary:
          type: string
        privateDetails:
          type: string

    # ── Procedure schemas ────────────────────────────────────────────────────
...
    ConductRecordResolveBody:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string
...