# lib/api-zod/src/generated/types/conductRecordCreate.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * C100 System API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConductRecordCreate {
  recordType: string;
  reportDate: string;
  /** @nullable */
  memberId?: number | null;
  /** @nullable */
  eventId?: number | null;
  summary: string;
  privateDetails?: string;
}