# lib/api-zod/src/generated/types/governanceDocumentCreate.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * C100 System API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GovernanceDocumentCreate {
  title: string;
  category: string;
  versionLabel: string;
  effectiveDate: string;
  /** @nullable */
  approvalDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  originalFilename?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  fileSizeBytes?: number | null;
  /** @nullable */
  storageKey?: string | null;
}